Camera2api added
Found that camera class is deprecated, and that camera2 is new class to replace it. Going to attempt to move forward with this camera2 implementation.

Next steps:
Captured image replaces preview
Retake picture button appears
OpenCV image processing on circles is displayed on captured image
Circle count is displayed with option to increase/decrease count (+50 button?)